refactor: split `main` into API-specific files
